+++ packages/icon/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [0.1.8](https://github.com/deepecom/deepui/compare/@deepui/icon@0.1.7...@deepui/icon@0.1.8) (2021-12-05)
+
+**Note:** Added Icons - SearchIcon, CheckIcon, ChevronDownIcon
+
+
+
 ## [0.1.7](https://github.com/deepecom/deepui/compare/@deepui/icon@0.1.6...@deepui/icon@0.1.7) (2021-08-28)
 
 **Note:** Version bump only for package @deepui/icon