+++ packages/selectbox/CHANGELOG.md
@@ -3,7 +3,11 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.1](https://github.com/deepecom/deepui/compare/@deepui/selectbox@0.1.1...@deepui/selectbox@0.1.2) (2021-10-08)
+## [0.1.3](https://github.com/deepecom/deepui/compare/@deepui/selectbox@0.1.2...@deepui/selectbox@0.1.3) (2021-12-05)
+
+**Note:** Updated Icons to not use heroicons and use deepIcon
+
+## [0.1.2](https://github.com/deepecom/deepui/compare/@deepui/selectbox@0.1.1...@deepui/selectbox@0.1.2) (2021-10-08)
 
 **Note:** Updated SearchAutoComplete to have focus and onBlur disabled for customised use